src/api-producer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
kafka-python==2.0.2
python-dotenv==1.0.1
pydantic==2.9.2
httpx>=0.27
